% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplot.R
\name{corrplot}
\alias{corrplot}
\title{Generates a correlation plot for the numeric variables in a dataframe,
and prints the correlation values.}
\usage{
corrplot(df, print_corr = FALSE, title = "Correlation Plot")
}
\arguments{
\item{df}{The dataframe containing the data to be plotted.}

\item{print_corr}{Whether or not to print the correlation values between variables}

\item{title}{Title of the plot}
}
\value{
A ggplot object.
}
\description{
Generates a correlation plot for the numeric variables in a dataframe,
and prints the correlation values.
}
\examples{
penguins_data <- palmerpenguins::penguins
corrplot(penguins_data, print_corr = FALSE, title = "Correlation chart")
}
