% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{Create a histogram with the magma color scheme}
\usage{
histogram(df, x, y)
}
\arguments{
\item{df}{A dataframe containing the variables for plotting}

\item{x}{Column name of the variable to be plotted on the x-axis}

\item{y}{An aggregation function in string version to be plotted on the y-axis.
The supported aggregation operations are: \link[='average', 'count',
           'distinct', 'max', 'mean', 'median', 'min', 'missing', 'product',
           'q1', 'q3', 'ci0', 'ci1', 'stderr', 'stdev', 'stdevp', 'sum', 
           'valid', 'values', 'variance', 'variancep']{'average', 'count', 'distinct', 'max', 'mean', 'median', 'min', 'missing', 'product', 'q1', 'q3', 'ci0', 'ci1', 'stderr', 'stdev', 'stdevp', 'sum', 'valid', 'values', 'variance', 'variancep'}}
}
\value{
A ggplot object.
}
\description{
Create a histogram with the magma color scheme
}
\examples{
histogram(mtcars, vs, 'min(mpg)')
histogram(mtcars, vs, 'count()')
}
