% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Create a scatterplot using the magma color scheme}
\usage{
scatterplot(
  df,
  x,
  y,
  c = "",
  t = "",
  o = 0.5,
  s = 50,
  xtitle = "",
  ytitle = "",
  ctitle = "",
  xzero = False,
  yzero = False,
  shapes = True
)
}
\arguments{
\item{df}{A dataframe to use for the scatterplot.}

\item{x}{Column-name of the numerical variable to be plotted on the x-axis}

\item{y}{Column-name of the numerical variable to be plotted on the y-axis}

\item{c}{Column-name of the categorical variable to color-code the data points
Default value is blank for cases when there is no categorical column}

\item{t}{Title of the plot. Default value is blank.
If not provided, title will be computed based on x, y and/or c}

\item{o}{Opacity of the data points}

\item{s}{Size of the data points}

\item{xtitle}{Title of the x-axis. Default value is blank.
If not provided, title will be proper case of the x axis column}

\item{ytitle}{Title of the x-axis. Default value is blank.
If not provided, title will be proper case of the y axis column}

\item{ctitle}{Title of the x-axis. Default value is blank.
If not provided, title will be proper case of the color column}

\item{xzero}{Scale the x-axis to start from 0 by specifying True
Default value is set to False}

\item{yzero}{Scale the x-axis to start from 0 by specifying True
Default value is set to False}

\item{shapes}{Assign the color column to the shape attribute of the plot if True
Default value is set to True}
}
\value{
A ggplot object.
}
\description{
Create a scatterplot using the magma color scheme
}
\examples{
>>> scatterplot(iris, sepalLength, sepalWidth, species,
                "Iris Sepal Length vs Sepal Width across Species",
                1.0, 50, "Sepal Length", "Sepal Width", "", False, False, True)
}
