% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot}
\alias{boxplot}
\title{Plot a boxplot with the magma color scheme and an option to facet.}
\usage{
boxplot(df, x, y, facet = FALSE)
}
\arguments{
\item{df}{Dataframe containing the variables for plotting}

\item{x}{Column name of the categorical variable to view the distribution of}

\item{y}{Column name containing the numerical variables to assign boxes to}

\item{facet}{Determines whether separate graphs will be created for each category}
}
\value{
a ggplot object
}
\description{
Plot a boxplot with the magma color scheme and an option to facet.
}
\examples{
df <- iris
boxplot(df, Species, Sepal.Length, facet = TRUE)

}
