% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplot.R
\name{corrplot}
\alias{corrplot}
\title{Create a correlation plot}
\usage{
corrplot(df, print_corr = FALSE, shape = "circle")
}
\arguments{
\item{df}{A dataframe to use for the correlation plot.}

\item{print_corr}{Whether or not to print the correlation values between variables}

\item{shape}{Shape to use in the plot}
}
\value{
A ggplot object.
}
\description{
Create a correlation plot
}
\examples{
corrplot(df, print_corr = FALSE, shape = "square")
}
