% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{Create a histogram with the magma color scheme}
\usage{
histogram(df, x, y)
}
\arguments{
\item{df}{A dataframe containing the variables for plotting}

\item{x}{Column name of the numeric variable to be plotted on the x-axis}

\item{y}{The quoted aggregation function to be plotted on the y-axis
and to be used for the fill color.
The input needs to be in the form of "..name..",
where name can be values from the following list,
with explanations in the parenthesis:
{count (number of points in bin),
density (density of points in bin, scaled to integrate to 1),
ncount (count, scaled to maximum of 1),
ndensity (density, scaled to maximum of 1),
width (widths of bins)}}
}
\value{
A ggplot object.
}
\description{
Create a histogram with the magma color scheme
}
