% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Create a scatterplot using the magma color scheme}
\usage{
scatterplot(
  df,
  x,
  y,
  c = NULL,
  t = "",
  o = 0.5,
  s = 3,
  xtitle = "",
  ytitle = "",
  ctitle = "",
  xzero = FALSE,
  yzero = FALSE,
  shapes = FALSE
)
}
\arguments{
\item{df}{A dataframe to use for the scatterplot.}

\item{x}{Column-name of the numerical or categorical variable to be plotted on the x-axis}

\item{y}{Column-name of the numerical or categorical variable to be plotted on the y-axis}

\item{c}{Column-name of the categorical variable to color-code the data points
Default value is blank for cases when there is no categorical column}

\item{t}{Title of the plot. Default value is blank.
If not provided, title will be computed based on x, y and/or c}

\item{o}{Opacity of the data points. Number between 0.1 and 1.0 (inclusive)}

\item{s}{Size of the data points. Number between 1 and 5 (inclusive)}

\item{xtitle}{Title of the x-axis. Default value is blank.
If not provided, title will be proper case of the x axis column}

\item{ytitle}{Title of the x-axis. Default value is blank.
If not provided, title will be proper case of the y axis column}

\item{ctitle}{Title of the x-axis. Default value is blank.
If not provided, title will be proper case of the color column}

\item{xzero}{Scale the x-axis to start from 0 by specifying TRUE
Default value is set to FALSE}

\item{yzero}{Scale the x-axis to start from 0 by specifying TRUE
Default value is set to FALSE}

\item{shapes}{Assign the color column to the shape attribute of the plot if TRUE
Default value is set to FALSE}
}
\value{
A ggplot object.
}
\description{
Create a scatterplot using the magma color scheme
}
\examples{
penguins_data <- palmerpenguins::penguins
scatterplot(penguins_data, bill_length_mm, flipper_length_mm, species,
              "Bill and Flipper length clusters by Species",
              0.5, 2.5, "Bill length (mm)", "Flipper length (mm)", "", FALSE, FALSE, TRUE)
}
